relexd skd duh d d thmp
Creating masthead and usage notes for graph, but there's a lot of files and stats to keep track of.

I'm introducing the plan for the file "quick_stats.txt"

A summary file that provides tabular per-file k-mer counts, with associated metadata

I'm also at a comfortable state of the edge list generation loops.

There's a lot to be done and I don't know how much to document vs trusting in the intuitive nature of the interface.

Still working on personal issues, 20-30h work.

Honest assessment of myself.